pyproject.toml
[tool.poetry]
name = "shoestore"
version = "0.1.0"
description = ""
authors = ["Cory Schlaudecker <cory.schlaudecker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
